Move token code into its own directory.
