+++ lib/index.js
@@ -4,7 +4,7 @@
 import * as batchVersions from './batchVersions.js';
 import * as documents from './documents.js';
 import * as entities from './entities.js';
-import * as tokens from './tokens.js';
+import * as tokens from './tokens/index.js';
 import './config.js';
 
 export {documents, entities, tokens, batchVersions};
