+++ lib/tokens/batches.js
@@ -2,11 +2,11 @@
  * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
 import * as base64url from 'base64url-universal';
-import * as batchVersions from './batchVersions.js';
+import * as batchVersions from '../batchVersions.js';
 import * as bedrock from '@bedrock/core';
 import * as database from '@bedrock/mongodb';
-import * as entities from './entities.js';
-import {create as _createToken} from './tokenFormat.js';
+import * as entities from '../entities.js';
+import {create as _createToken} from './format.js';
 import assert from 'assert-plus';
 import crypto from 'node:crypto';
 import pLimit from 'p-limit';
