+++ lib/tokens/format.js
@@ -2,7 +2,7 @@
  * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
 import * as base58 from 'base58-universal';
-import * as batchVersions from './batchVersions.js';
+import * as batchVersions from '../batchVersions.js';
 import * as bedrock from '@bedrock/core';
 import {
   decode as cborldDecode, encode as cborldEncode
