+++ lib/tokens/index.js
@@ -1,24 +1,24 @@
 /*!
  * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import * as batchVersions from './batchVersions.js';
-import * as documents from './documents.js';
-import {createTokens as _createTokens} from './tokenBatches.js';
+import * as batchVersions from '../batchVersions.js';
+import * as documents from '../documents.js';
+import {createTokens as _createTokens} from './batches.js';
 import assert from 'assert-plus';
 import {tokenizers} from '@bedrock/tokenizer';
 
 // expose public functions
-export {resolve, resolveToInternalId} from './resolveTokens.js';
-export {invalidateBatches as invalidateTokenBatches} from './tokenBatches.js';
+export {resolve, resolveToInternalId} from './resolve.js';
+export {invalidateBatches as invalidateTokenBatches} from './batches.js';
 
 // re-export functions for testing
 export {
-  getPairwiseToken as _getPairwiseToken,
-  upsertPairwiseToken as _upsertPairwiseToken
-} from './pairwiseTokens.js';
+  get as _getPairwiseToken,
+  upsert as _upsertPairwiseToken
+} from './pairwise.js';
 export {
   _claimTokens, getBatch as _getBatch, updateBatch as _updateBatch
-} from './tokenBatches.js';
+} from './batches.js';
 
 /**
  * Safely registers a `document` and creates tokens concurrently. Use this
