+++ lib/tokens/pairwise.js
@@ -29,7 +29,7 @@ bedrock.events.on('bedrock-mongodb.ready', async () => {
   // versions did not include `expires` in the records
 });
 
-export async function getPairwiseToken({
+export async function get({
   internalId, requester, explain = false
 } = {}) {
   const query = {
@@ -60,17 +60,17 @@ export async function getPairwiseToken({
 }
 
 // FIXME: accept `expires` for pairwise token
-export async function upsertPairwiseToken({internalId, requester}) {
+export async function upsert({internalId, requester}) {
   // FIXME: do actual upsert to update `expires` for pairwise token using
   // `$max` -- concurrently do a `get` to retrieve its actual value ... and
-  // then remove the `createPairwiseToken` thing ... ensure we do not
-  // retrieve nor overwrite a pairwise token if `requester` does not match
+  // then remove `_create` ... ensure we do not retrieve nor overwrite a
+  // pairwise token if `requester` does not match
 
   while(true) {
     // default to getting pairwise token first assuming that the common case
     // is that the requester has resolved a token for `internalId` before
     try {
-      const record = await getPairwiseToken({internalId, requester});
+      const record = await get({internalId, requester});
       return record;
     } catch(e) {
       if(e.name !== 'NotFoundError') {
@@ -80,7 +80,7 @@ export async function upsertPairwiseToken({internalId, requester}) {
     // create a pairwise token since one was not found
     try {
       // FIXME: pass `expires` for pairwise token?
-      const record = await _createPairwiseToken({internalId, requester});
+      const record = await _create({internalId, requester});
       return record;
     } catch(e) {
       if(e.name !== 'DuplicateError') {
@@ -92,7 +92,7 @@ export async function upsertPairwiseToken({internalId, requester}) {
 }
 
 // FIXME: accept `expires` for pairwise token
-async function _createPairwiseToken({internalId, requester}) {
+async function _create({internalId, requester}) {
   // generate pairwise token value
   const value = Buffer.from(await idGenerator.generate());
