+++ lib/tokens/resolve.js
@@ -3,16 +3,16 @@
  */
 import * as base64url from 'base64url-universal';
 import * as bedrock from '@bedrock/core';
-import * as entities from './entities.js';
+import * as entities from '../entities.js';
 import {
   getBatch as _getBatch,
   updateBatch as _updateBatch
-} from './tokenBatches.js';
+} from './batches.js';
 import {
-  getPairwiseToken as _getPairwiseToken,
-  upsertPairwiseToken as _upsertPairwiseToken
-} from './pairwiseTokens.js';
-import {parse as _parseToken} from './tokenFormat.js';
+  get as _getPairwiseToken,
+  upsert as _upsertPairwiseToken
+} from './pairwise.js';
+import {parse as _parseToken} from './format.js';
 import {Bitstring} from '@digitalbazaar/bitstring';
 
 const {util: {BedrockError}} = bedrock;
